Cleanup pagination bar component
