+++ ui/src/design-system/components/pagination-bar/page-button/page-button.module.scss
@@ -1,0 +1,7 @@
+@import 'design-system/variables/typography.scss';
+
+.pageButton {
+  @include paragraph-small();
+  font-weight: 600;
+  padding: 0 8px;
+}
